try gdc for ours

+++ Enzymes/deepinfomax_vae.py
@@ -213,7 +213,7 @@ class GcnInfomax(nn.Module):
                 x, edge_index, batch = data.x, data.edge_index, data.batch
 
 
-                node_mu, node_logvar, class_mu, class_logvar, entangledrep = self.encoder(x[:,:18], edge_index, batch)
+                node_mu, node_logvar, class_mu, class_logvar, entangledrep = self.encoder(x[:,:18].double(), edge_index, batch)
 
 
                 node_latent_embeddings = reparameterize(training=False, mu=node_mu, logvar=node_logvar)
@@ -389,7 +389,7 @@ if __name__ == '__main__':
 
                 optimizer.zero_grad()
                 #recon_loss, kl_class, kl_node = model(data.x[:,:18], data.edge_index, data.batch, data.num_graphs)
-                current_loss = model(data.x[:,:18], data.edge_index, data.batch, data.num_graphs)
+                current_loss = model(data.x[:,:18].double(), data.edge_index, data.batch, data.num_graphs)
                 '''recon_loss_all += recon_loss
                 kl_class_loss_all += kl_class
                 kl_node_loss_all += kl_node'''
